Update decoder and ocf_fsm
Now the decoder will accurately step through each
instruction, but it will do it cycle by cycle
instead of by macrostate. This is because the
transitions between macrostates are harder to
predict and require a bit more flexibility with
the decoder to be able to choose a transition
in the middle of an arbitrary macrostate for
when the macrostate has finished.