Changin our memory controller to work for writes
Reads give out the value on the next cycle when we
tell the memory we are reading, which is fine,
but that can't happen for writes. When we say
we are writing, the data is available now,
not on the next cc. So we have to use blocking
assignments instead of non-blocking to make these
writes work.