ADD A, r
This one got kind of tricky and some changes were necessary. The 8
bit alu now only outputs to the A-register. This change was made
because we need to be able to put an argument from the regfile
onto the databus in the same cycle that we load the A register.
This means that the 8-bit alu can't output its result to the A register
orelse we would have drive fight.

The other thing we needed to change is that the B register for the
ALU isn't permanent. For 4 cycle ops, we cant take the extra cycle
to load the TEMP register before using it. It is handy, however to
have it connected, so we put a mux in instead of removing it.

A new macro was added to reduce the complexity of using this feature
in the future. The macro is 8_ADD