LD_dd_nn_x fixes, LD_nn_x_dd implementation, LD_nn_x_HL
LD_dd_nn_x was broken for SP, that is fixed.

In order to implement the latter two instructions, we needed
to load MAR from the data bus instead of the addr bus. Those
instructions are now implemented and tested as well.

Suyash missed a few things in his test cases, so we will
have to go back and write tests for all of those instructions
to verify correctness.