INC r
This operation needed to be performed in a single cycle,
so we had to have the 8-bit alu have a p2p connection back
to the register file. This was done in the same way that
A has the p2p connection to do adds in one cycle.

We also had to distinguish between 8-bit incs and 16-bit
incs since they affect the flags differently. We had
to go back and change some old code to do this. At some
point we should go back and refactor old microcode to use
the macros so updates like this can be applied globally
by changing the macro.