Maskable Interrupts
Implement support for maskable interrupts wherein interrupts
are masked unless explicitly enabled. When we enter into an
interrupt handler, we automatically stack the state of the
interrupt mask and disable interrupts for the duration of the
handler. THe user can re-enable interrupts during an ISR in
order to create pre-emptable interrupts, although I am not
sure that this is used in the Sega Master System.