Remove blocking on scrolling so that characters display
There was a line in the bios that waits for synchronization before
displaying character patterns on the screen. Since we haven't
implemented scrolling... or reading from VRAM, this wont work.
SO we remove the line that does this to start debugging why
nothing shows up on the screen. Now we are displaying some stuff
on the screen, which is pretty cool.